Split LoadControl reflection logic
